Move result-keys.ts to /view/results
